+++ components/Feature/SewerPipeEditor.tsx
@@ -1,11 +1,11 @@
 import { EditorContainer } from './EditorContainer';
-import { OptionSelector } from './OptionSelector';
+import { RadioSelector } from './OptionSelector';
 
 export function SewerPipeEditor({ sewerPipeData, onChange }) {
   return (
     <EditorContainer>
       <h1 className='font-semibold'>Viemäriputkien tiedot</h1>
-      <OptionSelector
+      <RadioSelector
         label='Toteutustapa'
         labelKey='label'
         valueKey='id'